refactored backend structure, edited docker compose
