+++ backend/app/__main__.py
@@ -1,14 +1,14 @@
 import os
 
-from app import app
-from travel import travel
-from models import *
+from app.database import db
+from app.core import app
+from app.api import blueprints
 
 with app.app_context():
     db.create_all()
 
-app.register_blueprint(travel)
-
+for blueprint in blueprints:
+    app.register_blueprint(blueprint)
 
 if __name__ == "__main__":
     app.run(os.getenv("HOST", "0.0.0.0"), int(os.getenv("PORT", 8888)))
