+++ backend/app/api/__init__.py
@@ -1,0 +1,6 @@
+
+from .travel import api_travel
+
+blueprints = [
+    api_travel
+]
