+++ backend/app/database.py
@@ -1,0 +1,4 @@
+from flask_sqlalchemy import SQLAlchemy
+from app.core import app
+
+db = SQLAlchemy(app)
